Modularize plugin database tests
Separate bindings, query-channel, transaction, CloudSync lifecycle, credential, and shared test support from the production plugin facade without changing behavior.
