+++ plugins/db/src/lib.rs
@@ -279,1322 +279,4 @@ pub fn init_with_cloudsync<R: tauri::Runtime>(
 }
 
 #[cfg(test)]
-mod test {
-    use std::sync::{Arc, Mutex};
-    use std::time::Duration;
-
-    use anlg_db_reactive::QueryEventSink;
-    use serde_json::json;
-    use sqlx::Connection;
-    use tauri::ipc::{Channel, InvokeResponseBody};
-    use wiremock::{Mock, MockServer, Request, Respond, ResponseTemplate, matchers::path};
-
-    use super::*;
-
-    #[test]
-    fn export_types() {
-        const OUTPUT_FILE: &str = "./js/bindings.gen.ts";
-
-        make_specta_builder::<tauri::Wry>()
-            .export(
-                specta_typescript::Typescript::default()
-                    .formatter(specta_typescript::formatter::prettier)
-                    .bigint(specta_typescript::BigIntExportBehavior::Number),
-                OUTPUT_FILE,
-            )
-            .unwrap();
-
-        let content = std::fs::read_to_string(OUTPUT_FILE).unwrap();
-        std::fs::write(OUTPUT_FILE, format!("// @ts-nocheck\n{content}")).unwrap();
-    }
-
-    #[test]
-    fn default_permissions_exclude_generic_cloudsync_control() {
-        let permissions = include_str!("../permissions/default.toml");
-
-        assert!(permissions.contains("allow-configure-cloudsync-token"));
-        assert!(permissions.contains("allow-suspend-cloudsync-for-sign-out"));
-        assert!(!permissions.contains("allow-begin-cloudsync-activity"));
-        assert!(!permissions.contains("allow-end-cloudsync-activity"));
-        assert!(!permissions.contains("\"allow-configure-cloudsync\""));
-        assert!(!permissions.contains("allow-start-cloudsync"));
-        assert!(!permissions.contains("allow-sync-cloudsync-now"));
-    }
-
-    #[test]
-    fn default_permissions_include_legacy_migration_workflow() {
-        let permissions = include_str!("../permissions/default.toml");
-
-        for permission in [
-            "allow-get-legacy-cleanup-status",
-            "allow-get-legacy-import-report",
-            "allow-cleanup-legacy-files",
-            "allow-run-legacy-import",
-        ] {
-            assert!(permissions.contains(permission), "missing {permission}");
-        }
-    }
-
-    fn capture_channel() -> (Channel<QueryEvent>, Arc<Mutex<Vec<QueryEvent>>>) {
-        let events = Arc::new(Mutex::new(Vec::new()));
-        let captured = Arc::clone(&events);
-        let channel = Channel::new(move |body| {
-            let InvokeResponseBody::Json(payload) = body else {
-                return Ok(());
-            };
-            let event: QueryEvent =
-                serde_json::from_str(&payload).expect("channel payload should parse");
-            captured.lock().unwrap().push(event);
-            Ok(())
-        });
-        (channel, events)
-    }
-
-    async fn next_event(
-        events: &Arc<Mutex<Vec<QueryEvent>>>,
-        index: usize,
-    ) -> anyhow::Result<QueryEvent> {
-        tokio::time::timeout(Duration::from_secs(1), async {
-            loop {
-                if let Some(event) = events.lock().unwrap().get(index).cloned() {
-                    return event;
-                }
-                tokio::time::sleep(Duration::from_millis(10)).await;
-            }
-        })
-        .await
-        .map_err(anyhow::Error::from)
-    }
-
-    async fn setup_runtime_with_cloudsync(
-        cloudsync_enabled: bool,
-    ) -> (tempfile::TempDir, Arc<runtime::PluginDbRuntime>) {
-        let dir = tempfile::tempdir().unwrap();
-        let db_path = dir.path().join("app.db");
-        let db = anlg_db_core::Db::open(anlg_db_core::DbOpenOptions {
-            storage: anlg_db_core::DbStorage::Local(&db_path),
-            cloudsync_enabled,
-            journal_mode_wal: true,
-            foreign_keys: true,
-            max_connections: Some(4),
-        })
-        .await
-        .unwrap();
-        anlg_db_app::prepare_schema(&db).await.unwrap();
-        sqlx::query(
-            "UPDATE storage_migration_state
-             SET importer_version = ?, parity_verified = 1
-             WHERE id = 'legacy_v1'",
-        )
-        .bind(anlg_db_app::LEGACY_IMPORTER_VERSION)
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        (dir, Arc::new(runtime::PluginDbRuntime::new(Arc::new(db))))
-    }
-
-    async fn setup_runtime() -> (tempfile::TempDir, Arc<runtime::PluginDbRuntime>) {
-        setup_runtime_with_cloudsync(false).await
-    }
-
-    async fn setup_enabled_cloudsync_runtime() -> (tempfile::TempDir, Arc<runtime::PluginDbRuntime>)
-    {
-        let (dir, runtime) = setup_runtime_with_cloudsync(true).await;
-        let recovery_key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap();
-        runtime
-            .set_e2ee_recovery_key("user-a", &recovery_key)
-            .unwrap();
-        (dir, runtime)
-    }
-
-    #[derive(Clone, Default)]
-    struct WitnessResponder {
-        events: Arc<Mutex<Vec<serde_json::Value>>>,
-    }
-
-    impl Respond for WitnessResponder {
-        fn respond(&self, request: &Request) -> ResponseTemplate {
-            if request.method == wiremock::http::Method::POST {
-                let body: serde_json::Value = request.body_json().unwrap();
-                let mut events = self.events.lock().unwrap();
-                for event in body["events"].as_array().unwrap() {
-                    let duplicate = events.iter().any(|existing| {
-                        existing["recordId"] == event["recordId"]
-                            && existing["payloadHash"] == event["payloadHash"]
-                    });
-                    if !duplicate {
-                        let mut event = event.clone();
-                        event["sequence"] = json!(events.len() + 1);
-                        events.push(event);
-                    }
-                }
-                return ResponseTemplate::new(200).set_body_json(json!({
-                    "initializedAt": "2026-07-17T00:00:00Z",
-                    "headSequence": events.len(),
-                }));
-            }
-
-            let query = request
-                .url
-                .query_pairs()
-                .collect::<std::collections::HashMap<_, _>>();
-            let after = query
-                .get("afterSequence")
-                .and_then(|value| value.parse::<u64>().ok())
-                .unwrap_or(0);
-            let events = self.events.lock().unwrap();
-            let head = u64::try_from(events.len()).unwrap();
-            let through = query
-                .get("throughSequence")
-                .and_then(|value| value.parse::<u64>().ok())
-                .unwrap_or(head);
-            let page = events
-                .iter()
-                .filter(|event| {
-                    event["sequence"].as_u64().unwrap() > after
-                        && event["sequence"].as_u64().unwrap() <= through
-                })
-                .take(3)
-                .cloned()
-                .collect::<Vec<_>>();
-            let next = page
-                .last()
-                .and_then(|event| event["sequence"].as_u64())
-                .unwrap_or(after);
-            ResponseTemplate::new(200).set_body_json(json!({
-                "initialized": true,
-                "initializedAt": "2026-07-17T00:00:00Z",
-                "headSequence": head,
-                "throughSequence": through,
-                "nextAfterSequence": next,
-                "events": page,
-            }))
-        }
-    }
-
-    async fn setup_witness(workspace_id: &str) -> (MockServer, CloudsyncE2eeWitness) {
-        let server = MockServer::start().await;
-        Mock::given(path(format!("/sync/e2ee/witness/{workspace_id}")))
-            .respond_with(WitnessResponder::default())
-            .mount(&server)
-            .await;
-        let config = CloudsyncE2eeWitness {
-            endpoint: format!("{}/sync/e2ee/witness/{workspace_id}", server.uri()),
-            access_token: "access-token".to_string(),
-        };
-        (server, config)
-    }
-
-    fn unreachable_witness(workspace_id: &str) -> CloudsyncE2eeWitness {
-        CloudsyncE2eeWitness {
-            endpoint: format!("http://127.0.0.1:9/sync/e2ee/witness/{workspace_id}"),
-            access_token: "access-token".to_string(),
-        }
-    }
-
-    async fn setup_unmigrated_runtime_with_max_connections(
-        max_connections: u32,
-    ) -> (tempfile::TempDir, Arc<runtime::PluginDbRuntime>) {
-        let dir = tempfile::tempdir().unwrap();
-        let db_path = dir.path().join("app.db");
-        let db = anlg_db_core::Db::open(anlg_db_core::DbOpenOptions {
-            storage: anlg_db_core::DbStorage::Local(&db_path),
-            cloudsync_enabled: false,
-            journal_mode_wal: true,
-            foreign_keys: true,
-            max_connections: Some(max_connections),
-        })
-        .await
-        .unwrap();
-
-        (dir, Arc::new(runtime::PluginDbRuntime::new(Arc::new(db))))
-    }
-
-    async fn setup_unmigrated_runtime() -> (tempfile::TempDir, Arc<runtime::PluginDbRuntime>) {
-        setup_unmigrated_runtime_with_max_connections(4).await
-    }
-
-    #[tokio::test]
-    async fn sign_out_suspend_command_preserves_activity_leases() {
-        let (_dir, runtime) = setup_runtime().await;
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        let app = tauri::test::mock_builder()
-            .manage(Arc::clone(&runtime))
-            .build(tauri::test::mock_context(tauri::test::noop_assets()))
-            .unwrap();
-
-        commands::suspend_cloudsync_for_sign_out(app.state())
-            .await
-            .unwrap();
-
-        let status = runtime.cloudsync_status().await.unwrap();
-        assert_eq!(status["activity_paused"], true);
-        assert_eq!(status["deferred_for_capture"], true);
-    }
-
-    #[cfg(any(
-        all(target_os = "macos", target_arch = "aarch64"),
-        all(target_os = "macos", target_arch = "x86_64"),
-        all(target_os = "linux", target_env = "gnu", target_arch = "aarch64"),
-        all(target_os = "linux", target_env = "gnu", target_arch = "x86_64"),
-        all(target_os = "linux", target_env = "musl", target_arch = "aarch64"),
-        all(target_os = "linux", target_env = "musl", target_arch = "x86_64"),
-        all(target_os = "windows", target_arch = "x86_64"),
-    ))]
-    #[tokio::test]
-    async fn sign_out_suspend_command_defers_busy_pool_teardown() {
-        let dir = tempfile::tempdir().unwrap();
-        let db_path = dir.path().join("app.db");
-        let db = anlg_db_core::Db::open(anlg_db_core::DbOpenOptions {
-            storage: anlg_db_core::DbStorage::Local(&db_path),
-            cloudsync_enabled: true,
-            journal_mode_wal: true,
-            foreign_keys: true,
-            max_connections: Some(2),
-        })
-        .await
-        .unwrap();
-        anlg_db_app::prepare_schema(&db).await.unwrap();
-        db.cloudsync_init("sessions", None, None).await.unwrap();
-        let runtime = Arc::new(runtime::PluginDbRuntime::new(Arc::new(db)));
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        let held_connection = runtime.pool().acquire().await.unwrap();
-        let app = tauri::test::mock_builder()
-            .manage(Arc::clone(&runtime))
-            .build(tauri::test::mock_context(tauri::test::noop_assets()))
-            .unwrap();
-
-        tokio::time::timeout(
-            Duration::from_millis(1_500),
-            commands::suspend_cloudsync_for_sign_out(app.state()),
-        )
-        .await
-        .expect("sign-out suspension exceeded its pool-drain deadline")
-        .unwrap();
-
-        drop(held_connection);
-        let mut replacement =
-            tokio::time::timeout(Duration::from_millis(1_500), runtime.pool().acquire())
-                .await
-                .expect("pool did not open a replacement after deferred teardown")
-                .unwrap();
-        replacement.return_to_pool().await;
-        let status = runtime.cloudsync_status().await.unwrap();
-        assert_eq!(status["activity_paused"], true);
-        assert_eq!(status["deferred_for_capture"], true);
-
-        runtime.suspend_cloudsync().await.unwrap();
-        tokio::time::timeout(
-            Duration::from_millis(1_500),
-            sqlx::query(
-                "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-                 VALUES ('after-sign-out-retry', '', '', 'Note')",
-            )
-            .execute(runtime.pool()),
-        )
-        .await
-        .expect("local write could not reuse the pool after deferred teardown retry")
-        .unwrap();
-    }
-
-    #[tokio::test]
-    async fn query_event_channel_sends_result_payload() {
-        let (channel, events) = capture_channel();
-        let sink = runtime::QueryEventChannel::new(channel);
-
-        sink.send_result(vec![json!({ "id": "note-1" })]).unwrap();
-
-        let event = next_event(&events, 0).await.unwrap();
-        assert_eq!(event, QueryEvent::Result(vec![json!({ "id": "note-1" })]));
-    }
-
-    #[tokio::test]
-    async fn query_event_channel_sends_error_payload() {
-        let (channel, events) = capture_channel();
-        let sink = runtime::QueryEventChannel::new(channel);
-
-        sink.send_error("boom".to_string()).unwrap();
-
-        let event = next_event(&events, 0).await.unwrap();
-        assert_eq!(event, QueryEvent::Error("boom".to_string()));
-    }
-
-    #[test]
-    fn query_event_serializes_with_tagged_shape() {
-        let result =
-            serde_json::to_value(QueryEvent::Result(vec![json!({ "id": "note-1" })])).unwrap();
-        let error = serde_json::to_value(QueryEvent::Error("boom".to_string())).unwrap();
-
-        assert_eq!(
-            result,
-            json!({ "event": "result", "data": [{ "id": "note-1" }] })
-        );
-        assert_eq!(error, json!({ "event": "error", "data": "boom" }));
-    }
-
-    #[tokio::test]
-    async fn subscribe_sends_initial_result_through_channel() {
-        let (_dir, runtime) = setup_runtime().await;
-        let (channel, events) = capture_channel();
-
-        runtime
-            .subscribe(
-                "SELECT id, title FROM templates WHERE id = 'missing-template' ORDER BY id"
-                    .to_string(),
-                vec![],
-                runtime::QueryEventChannel::new(channel),
-            )
-            .await
-            .unwrap();
-
-        let event = next_event(&events, 0).await.unwrap();
-        assert_eq!(event, QueryEvent::Result(Vec::new()));
-    }
-
-    #[tokio::test]
-    async fn execute_proxy_applies_app_schema_before_run() {
-        let (_dir, runtime) = setup_unmigrated_runtime().await;
-
-        runtime
-            .execute_proxy(
-                "INSERT INTO templates (id, title) VALUES (?, ?)".to_string(),
-                vec![json!("template-1"), json!("Template 1")],
-                anlg_db_execute::ProxyQueryMethod::Run,
-            )
-            .await
-            .unwrap();
-
-        let rows = runtime
-            .execute(
-                "SELECT id, title FROM templates WHERE id = ?".to_string(),
-                vec![json!("template-1")],
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(
-            rows,
-            vec![json!({
-                "id": "template-1",
-                "title": "Template 1",
-            })]
-        );
-    }
-
-    #[tokio::test]
-    async fn execute_transaction_commits_every_statement_atomically() {
-        let (_dir, runtime) = setup_unmigrated_runtime().await;
-
-        let rows_affected = runtime
-            .execute_transaction(vec![
-                TransactionStatement {
-                    sql: "INSERT INTO templates (id, title) VALUES (?, ?)".to_string(),
-                    params: vec![json!("template-1"), json!("Template 1")],
-                    expected_rows_affected: None,
-                },
-                TransactionStatement {
-                    sql: "INSERT INTO templates (id, title) VALUES (?, ?)".to_string(),
-                    params: vec![json!("template-2"), json!("Template 2")],
-                    expected_rows_affected: None,
-                },
-            ])
-            .await
-            .unwrap();
-
-        assert_eq!(rows_affected, vec![1, 1]);
-
-        let rows = runtime
-            .execute(
-                "SELECT id FROM templates WHERE id IN (?, ?) ORDER BY id".to_string(),
-                vec![json!("template-1"), json!("template-2")],
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(
-            rows,
-            vec![json!({ "id": "template-1" }), json!({ "id": "template-2" })]
-        );
-    }
-
-    #[tokio::test]
-    async fn execute_transaction_rolls_back_when_a_statement_fails() {
-        let (_dir, runtime) = setup_runtime().await;
-
-        let result = runtime
-            .execute_transaction(vec![
-                TransactionStatement {
-                    sql: "INSERT INTO templates (id, title) VALUES (?, ?)".to_string(),
-                    params: vec![json!("template-rollback"), json!("Rollback")],
-                    expected_rows_affected: None,
-                },
-                TransactionStatement {
-                    sql: "INSERT INTO missing_table (id) VALUES (?)".to_string(),
-                    params: vec![json!("fail")],
-                    expected_rows_affected: None,
-                },
-            ])
-            .await;
-
-        assert!(result.is_err());
-        let rows = runtime
-            .execute(
-                "SELECT id FROM templates WHERE id = ?".to_string(),
-                vec![json!("template-rollback")],
-            )
-            .await
-            .unwrap();
-        assert!(rows.is_empty());
-    }
-
-    #[tokio::test]
-    async fn execute_transaction_rolls_back_when_affected_rows_do_not_match() {
-        let (_dir, runtime) = setup_runtime().await;
-
-        let result = runtime
-            .execute_transaction(vec![
-                TransactionStatement {
-                    sql: "INSERT INTO templates (id, title) VALUES (?, ?)".to_string(),
-                    params: vec![json!("template-rollback"), json!("Rollback")],
-                    expected_rows_affected: Some(1),
-                },
-                TransactionStatement {
-                    sql: "UPDATE templates SET title = ? WHERE id = ?".to_string(),
-                    params: vec![json!("Missing"), json!("missing-template")],
-                    expected_rows_affected: Some(1),
-                },
-            ])
-            .await;
-
-        assert!(
-            result
-                .unwrap_err()
-                .to_string()
-                .contains("statement 1 affected 0 rows; expected 1")
-        );
-        let rows = runtime
-            .execute(
-                "SELECT id FROM templates WHERE id = ?".to_string(),
-                vec![json!("template-rollback")],
-            )
-            .await
-            .unwrap();
-        assert!(rows.is_empty());
-    }
-
-    #[tokio::test]
-    async fn cancelled_transaction_returns_a_clean_connection_to_the_pool() {
-        let (_dir, runtime) = setup_unmigrated_runtime_with_max_connections(1).await;
-        sqlx::query("CREATE TEMP TABLE pooled_connection_marker (id INTEGER PRIMARY KEY)")
-            .execute(runtime.pool())
-            .await
-            .unwrap();
-
-        runtime.pause_next_transaction_after_begin();
-        let transaction_runtime = Arc::clone(&runtime);
-        let transaction_task = tokio::spawn(async move {
-            transaction_runtime
-                .execute_transaction(vec![TransactionStatement {
-                    sql: "INSERT INTO templates (id, title) VALUES (?, ?)".to_string(),
-                    params: vec![json!("cancelled"), json!("Cancelled")],
-                    expected_rows_affected: Some(1),
-                }])
-                .await
-        });
-
-        runtime.wait_for_transaction_after_begin().await;
-        transaction_task.abort();
-        assert!(transaction_task.await.unwrap_err().is_cancelled());
-
-        let mut connection = tokio::time::timeout(Duration::from_secs(2), runtime.pool().acquire())
-            .await
-            .expect("timed out reacquiring the pooled connection")
-            .unwrap();
-        assert!(!connection.is_in_transaction());
-        let marker_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM pooled_connection_marker")
-            .fetch_one(&mut *connection)
-            .await
-            .expect("the original pooled connection should be reused");
-        assert_eq!(marker_count, 0);
-
-        let mut transaction = connection
-            .begin_with("BEGIN IMMEDIATE")
-            .await
-            .expect("a new immediate transaction should start");
-        sqlx::query("INSERT INTO templates (id, title) VALUES (?, ?)")
-            .bind("after-cancellation")
-            .bind("After cancellation")
-            .execute(&mut *transaction)
-            .await
-            .expect("a local write should succeed after cancellation");
-        transaction.commit().await.unwrap();
-        drop(connection);
-
-        let rows = runtime
-            .execute(
-                "SELECT id FROM templates WHERE id = ?".to_string(),
-                vec![json!("after-cancellation")],
-            )
-            .await
-            .unwrap();
-        assert_eq!(rows, vec![json!({ "id": "after-cancellation" })]);
-    }
-
-    #[tokio::test]
-    async fn open_memory_app_db_subscribe_sees_app_schema() {
-        let db = runtime::open_app_db(None).await.unwrap();
-        let runtime = runtime::PluginDbRuntime::new(Arc::new(db));
-        let (channel, events) = capture_channel();
-
-        let registration = runtime
-            .subscribe(
-                "SELECT id, title FROM templates ORDER BY id".to_string(),
-                vec![],
-                runtime::QueryEventChannel::new(channel),
-            )
-            .await
-            .unwrap();
-
-        assert!(matches!(
-            registration.analysis,
-            anlg_db_reactive::DependencyAnalysis::Reactive { .. }
-        ));
-
-        let event = next_event(&events, 0).await.unwrap();
-        assert!(matches!(event, QueryEvent::Result(rows) if !rows.is_empty()));
-    }
-
-    #[tokio::test]
-    async fn cloudsync_transport_stays_inert_until_configured() {
-        let (_dir, runtime) = setup_runtime().await;
-
-        let status = runtime.cloudsync_status().await.unwrap();
-        assert_eq!(status["cloudsync_enabled"], false);
-        assert_eq!(status["configured"], false);
-
-        runtime
-            .configure_cloudsync(
-                serde_json::json!({
-                    "connection_string": "managed-database-id",
-                    "auth": { "type": "token", "token": "test-token" },
-                    "tables": anlg_db_app::cloudsync_table_registry(),
-                    "sync_interval_ms": 30_000,
-                    "wait_ms": 5_000,
-                    "max_retries": 3
-                })
-                .to_string(),
-            )
-            .await
-            .unwrap();
-        runtime.start_cloudsync().await.unwrap();
-
-        let status = runtime.cloudsync_status().await.unwrap();
-        assert_eq!(status["configured"], true);
-        assert_eq!(status["running"], false);
-        assert_eq!(status["network_initialized"], false);
-
-        runtime.logout_cloudsync(false).await.unwrap();
-        assert_eq!(
-            runtime.cloudsync_status().await.unwrap()["configured"],
-            false
-        );
-    }
-
-    #[tokio::test]
-    async fn cloudsync_waits_for_legacy_migration_verification() {
-        let (_dir, runtime) = setup_enabled_cloudsync_runtime().await;
-        sqlx::query(
-            "UPDATE storage_migration_state
-             SET parity_verified = 0, last_error = 'completed_with_issues'
-             WHERE id = 'legacy_v1'",
-        )
-        .execute(runtime.pool())
-        .await
-        .unwrap();
-
-        let error = runtime
-            .configure_cloudsync_token(
-                "managed-database-id".to_string(),
-                "token".to_string(),
-                "user-a".to_string(),
-                unreachable_witness("user-a"),
-            )
-            .await
-            .unwrap_err();
-
-        assert!(matches!(&error, crate::Error::Io(_)));
-        assert!(
-            error
-                .to_string()
-                .contains("migration needs attention before CloudSync can start")
-        );
-        assert!(runtime.start_cloudsync().await.is_err());
-        assert!(runtime.sync_cloudsync_now().await.is_err());
-        assert_eq!(
-            runtime.cloudsync_status().await.unwrap()["configured"],
-            false
-        );
-    }
-
-    #[tokio::test]
-    async fn account_binding_is_durable_without_rekeying_local_rows() {
-        let (_dir, runtime) = setup_runtime().await;
-        let local_workspace = anlg_db_app::ensure_cloudsync_workspace_binding(runtime.pool())
-            .await
-            .unwrap();
-        sqlx::query(
-            "INSERT INTO humans (id, workspace_id, owner_user_id, name)
-             VALUES (?, ?, ?, 'Local user')",
-        )
-        .bind(&local_workspace)
-        .bind(&local_workspace)
-        .bind(&local_workspace)
-        .execute(runtime.pool())
-        .await
-        .unwrap();
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session', ?, ?, 'Session')",
-        )
-        .bind(&local_workspace)
-        .bind(&local_workspace)
-        .execute(runtime.pool())
-        .await
-        .unwrap();
-
-        assert!(
-            runtime
-                .bind_cloudsync_account("user-a".to_string())
-                .await
-                .unwrap()
-        );
-
-        let binding: (String, String) = sqlx::query_as(
-            "SELECT json_extract(value_json, '$.workspace_id'),
-                    json_extract(value_json, '$.account_user_id')
-             FROM app_settings WHERE id = 'cloudsync_workspace_binding'",
-        )
-        .fetch_one(runtime.pool())
-        .await
-        .unwrap();
-        let human: (String, String, String) = sqlx::query_as(
-            "SELECT id, workspace_id, owner_user_id FROM humans WHERE name = 'Local user'",
-        )
-        .fetch_one(runtime.pool())
-        .await
-        .unwrap();
-        let session: (String, String) =
-            sqlx::query_as("SELECT workspace_id, owner_user_id FROM sessions WHERE id = 'session'")
-                .fetch_one(runtime.pool())
-                .await
-                .unwrap();
-
-        assert_eq!(binding, (local_workspace.clone(), "user-a".to_string()));
-        assert_eq!(
-            human,
-            (
-                local_workspace.clone(),
-                local_workspace.clone(),
-                local_workspace.clone(),
-            )
-        );
-        assert_eq!(session, (local_workspace.clone(), local_workspace));
-        assert!(
-            !runtime
-                .bind_cloudsync_account("user-b".to_string())
-                .await
-                .unwrap()
-        );
-        assert_eq!(
-            runtime.cloudsync_status().await.unwrap()["configured"],
-            false
-        );
-    }
-
-    #[tokio::test]
-    async fn token_configuration_rejects_local_only_runtime_before_rekeying() {
-        let (_dir, runtime) = setup_runtime().await;
-        let local_workspace = anlg_db_app::ensure_cloudsync_workspace_binding(runtime.pool())
-            .await
-            .unwrap();
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session', ?, ?, 'Session')",
-        )
-        .bind(&local_workspace)
-        .bind(&local_workspace)
-        .execute(runtime.pool())
-        .await
-        .unwrap();
-
-        let error = runtime
-            .configure_cloudsync_token(
-                "managed-database-id".to_string(),
-                "token".to_string(),
-                "user-a".to_string(),
-                unreachable_witness("user-a"),
-            )
-            .await
-            .unwrap_err();
-
-        let binding: (String, Option<String>) = sqlx::query_as(
-            "SELECT json_extract(value_json, '$.workspace_id'),
-                    json_extract(value_json, '$.account_user_id')
-             FROM app_settings WHERE id = 'cloudsync_workspace_binding'",
-        )
-        .fetch_one(runtime.pool())
-        .await
-        .unwrap();
-        let session: (String, String) =
-            sqlx::query_as("SELECT workspace_id, owner_user_id FROM sessions WHERE id = 'session'")
-                .fetch_one(runtime.pool())
-                .await
-                .unwrap();
-
-        assert!(matches!(
-            error,
-            crate::Error::Cloudsync(anlg_db_core::CloudsyncRuntimeError::Unavailable)
-        ));
-        assert_eq!(binding, (local_workspace.clone(), None));
-        assert_eq!(session, (local_workspace.clone(), local_workspace));
-        assert_eq!(
-            runtime.cloudsync_status().await.unwrap()["configured"],
-            false
-        );
-    }
-
-    #[tokio::test]
-    async fn invalid_projection_does_not_claim_an_unbound_database() {
-        let (_dir, runtime) = setup_enabled_cloudsync_runtime().await;
-        sqlx::query("DELETE FROM app_settings WHERE id = 'cloudsync_workspace_binding'")
-            .execute(runtime.pool())
-            .await
-            .unwrap();
-        sqlx::query("INSERT INTO sessions (id, title) VALUES ('session', 'Session')")
-            .execute(runtime.pool())
-            .await
-            .unwrap();
-
-        let error = runtime
-            .configure_cloudsync_token_with_projection(
-                "managed-database-id".to_string(),
-                "token".to_string(),
-                "user-a".to_string(),
-                Some(anlg_db_app::CloudsyncWorkspaceProjection {
-                    account_user_id: "user-a".to_string(),
-                    personal_workspace_id: "user-a".to_string(),
-                    workspaces: vec![],
-                }),
-                unreachable_witness("user-a"),
-            )
-            .await
-            .unwrap_err();
-
-        let binding_count: i64 = sqlx::query_scalar(
-            "SELECT COUNT(*) FROM app_settings WHERE id = 'cloudsync_workspace_binding'",
-        )
-        .fetch_one(runtime.pool())
-        .await
-        .unwrap();
-        let session: (String, String) =
-            sqlx::query_as("SELECT workspace_id, owner_user_id FROM sessions WHERE id = 'session'")
-                .fetch_one(runtime.pool())
-                .await
-                .unwrap();
-
-        assert!(matches!(
-            error,
-            crate::Error::CloudsyncWorkspace(
-                anlg_db_app::CloudsyncWorkspaceError::InvalidWorkspaceProjection
-            )
-        ));
-        assert_eq!(binding_count, 0);
-        assert_eq!(session, (String::new(), String::new()));
-    }
-
-    #[tokio::test]
-    async fn token_configuration_claims_workspace_and_can_be_suspended() {
-        let (_dir, runtime) = setup_enabled_cloudsync_runtime().await;
-        let (_witness_server, witness) = setup_witness("user-a").await;
-        let local_workspace = anlg_db_app::ensure_cloudsync_workspace_binding(runtime.pool())
-            .await
-            .unwrap();
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session', ?, ?, 'Session')",
-        )
-        .bind(&local_workspace)
-        .bind(&local_workspace)
-        .execute(runtime.pool())
-        .await
-        .unwrap();
-
-        assert!(
-            runtime
-                .bind_cloudsync_account("user-a".to_string())
-                .await
-                .unwrap()
-        );
-
-        assert_eq!(
-            runtime
-                .configure_cloudsync_token(
-                    "managed-database-id".to_string(),
-                    "token".to_string(),
-                    "user-a".to_string(),
-                    witness,
-                )
-                .await
-                .unwrap(),
-            CloudsyncTokenConfigurationResult::Configured
-        );
-        runtime.suspend_cloudsync().await.unwrap();
-
-        let session: (String, String) =
-            sqlx::query_as("SELECT workspace_id, owner_user_id FROM sessions WHERE id = 'session'")
-                .fetch_one(runtime.pool())
-                .await
-                .unwrap();
-        let binding: (String, String) = sqlx::query_as(
-            "SELECT json_extract(value_json, '$.workspace_id'),
-                    json_extract(value_json, '$.account_user_id')
-             FROM app_settings WHERE id = 'cloudsync_workspace_binding'",
-        )
-        .fetch_one(runtime.pool())
-        .await
-        .unwrap();
-
-        assert_eq!(session, ("user-a".to_string(), "user-a".to_string()));
-        assert_eq!(binding, ("user-a".to_string(), "user-a".to_string()));
-    }
-
-    #[tokio::test]
-    async fn token_refresh_restarts_pending_full_resync_with_new_credentials() {
-        let (_dir, runtime) = setup_enabled_cloudsync_runtime().await;
-        let (_witness_server, witness) = setup_witness("user-a").await;
-        let local_workspace = anlg_db_app::ensure_cloudsync_workspace_binding(runtime.pool())
-            .await
-            .unwrap();
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session', ?, ?, 'Session')",
-        )
-        .bind(&local_workspace)
-        .bind(&local_workspace)
-        .execute(runtime.pool())
-        .await
-        .unwrap();
-        assert!(
-            runtime
-                .bind_cloudsync_account("user-a".to_string())
-                .await
-                .unwrap()
-        );
-        let projection = anlg_db_app::CloudsyncWorkspaceProjection {
-            account_user_id: "user-a".to_string(),
-            personal_workspace_id: "user-a".to_string(),
-            workspaces: vec![anlg_db_app::CloudsyncWorkspaceProjectionEntry {
-                id: "user-a".to_string(),
-                owner_user_id: "user-a".to_string(),
-                kind: "personal".to_string(),
-                name: "Personal".to_string(),
-                membership_id: "membership-personal".to_string(),
-                role: "owner".to_string(),
-                membership_created_at: "2026-07-01T01:00:00Z".to_string(),
-                membership_updated_at: "2026-07-16T01:00:00Z".to_string(),
-                created_at: "2026-07-01T00:00:00Z".to_string(),
-                updated_at: "2026-07-16T00:00:00Z".to_string(),
-            }],
-        };
-
-        assert_eq!(
-            runtime
-                .configure_cloudsync_token_with_projection(
-                    "managed-database-id".to_string(),
-                    "initial-token".to_string(),
-                    "user-a".to_string(),
-                    Some(projection.clone()),
-                    witness.clone(),
-                )
-                .await
-                .unwrap(),
-            CloudsyncTokenConfigurationResult::Configured
-        );
-        let (generation, auth) = runtime.cloudsync_full_resync_task_snapshot().await.unwrap();
-        assert_eq!(
-            auth,
-            anlg_db_core::CloudsyncAuth::Token {
-                token: "initial-token".to_string(),
-            }
-        );
-
-        assert_eq!(
-            runtime
-                .configure_cloudsync_token_with_projection(
-                    "managed-database-id".to_string(),
-                    "refreshed-token".to_string(),
-                    "user-a".to_string(),
-                    Some(projection),
-                    witness,
-                )
-                .await
-                .unwrap(),
-            CloudsyncTokenConfigurationResult::Configured
-        );
-        let (refreshed_generation, refreshed_auth) =
-            runtime.cloudsync_full_resync_task_snapshot().await.unwrap();
-        assert_eq!(refreshed_generation, generation);
-        assert_eq!(
-            refreshed_auth,
-            anlg_db_core::CloudsyncAuth::Token {
-                token: "refreshed-token".to_string(),
-            }
-        );
-
-        runtime.suspend_cloudsync().await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn token_configuration_projects_server_workspaces_after_account_claim() {
-        let (_dir, runtime) = setup_enabled_cloudsync_runtime().await;
-        let (_witness_server, witness) = setup_witness("user-a").await;
-        assert!(
-            runtime
-                .bind_cloudsync_account("user-a".to_string())
-                .await
-                .unwrap()
-        );
-
-        let projection = anlg_db_app::CloudsyncWorkspaceProjection {
-            account_user_id: "user-a".to_string(),
-            personal_workspace_id: "user-a".to_string(),
-            workspaces: vec![
-                anlg_db_app::CloudsyncWorkspaceProjectionEntry {
-                    id: "user-a".to_string(),
-                    owner_user_id: "user-a".to_string(),
-                    kind: "personal".to_string(),
-                    name: "Personal".to_string(),
-                    membership_id: "membership-personal".to_string(),
-                    role: "owner".to_string(),
-                    membership_created_at: "2026-07-01T01:00:00Z".to_string(),
-                    membership_updated_at: "2026-07-16T01:00:00Z".to_string(),
-                    created_at: "2026-07-01T00:00:00Z".to_string(),
-                    updated_at: "2026-07-16T00:00:00Z".to_string(),
-                },
-                anlg_db_app::CloudsyncWorkspaceProjectionEntry {
-                    id: "workspace-shared".to_string(),
-                    owner_user_id: "user-b".to_string(),
-                    kind: "shared".to_string(),
-                    name: "Shared".to_string(),
-                    membership_id: "membership-shared".to_string(),
-                    role: "member".to_string(),
-                    membership_created_at: "2026-07-02T01:00:00Z".to_string(),
-                    membership_updated_at: "2026-07-15T01:00:00Z".to_string(),
-                    created_at: "2026-07-02T00:00:00Z".to_string(),
-                    updated_at: "2026-07-15T00:00:00Z".to_string(),
-                },
-            ],
-        };
-
-        assert_eq!(
-            runtime
-                .configure_cloudsync_token_with_projection(
-                    "managed-database-id".to_string(),
-                    "token".to_string(),
-                    "user-a".to_string(),
-                    Some(projection),
-                    witness,
-                )
-                .await
-                .unwrap(),
-            CloudsyncTokenConfigurationResult::Configured
-        );
-        runtime.suspend_cloudsync().await.unwrap();
-
-        let workspaces: Vec<(String, String)> =
-            sqlx::query_as("SELECT id, name FROM workspaces ORDER BY id")
-                .fetch_all(runtime.pool())
-                .await
-                .unwrap();
-        let memberships: Vec<(String, String, String)> = sqlx::query_as(
-            "SELECT workspace_id, user_id, role FROM workspace_memberships ORDER BY workspace_id",
-        )
-        .fetch_all(runtime.pool())
-        .await
-        .unwrap();
-        let writable_workspace_ids: Vec<String> = sqlx::query_scalar(
-            "SELECT allowed_workspace_id
-             FROM cloudsync_writable_workspaces
-             ORDER BY allowed_workspace_id",
-        )
-        .fetch_all(runtime.pool())
-        .await
-        .unwrap();
-
-        assert_eq!(
-            workspaces,
-            vec![
-                ("user-a".to_string(), "Personal".to_string()),
-                ("workspace-shared".to_string(), "Shared".to_string()),
-            ]
-        );
-        assert_eq!(
-            memberships,
-            vec![
-                (
-                    "user-a".to_string(),
-                    "user-a".to_string(),
-                    "owner".to_string(),
-                ),
-                (
-                    "workspace-shared".to_string(),
-                    "user-a".to_string(),
-                    "member".to_string(),
-                ),
-            ]
-        );
-        assert_eq!(writable_workspace_ids, vec!["user-a".to_string()]);
-        assert!(
-            anlg_db_app::cloudsync_write_filter_installed(runtime.pool(), "user-a")
-                .await
-                .unwrap()
-        );
-        assert!(runtime.cloudsync_write_filters_match().await.unwrap());
-    }
-
-    #[tokio::test]
-    async fn token_configuration_account_mismatch_preserves_workspace_projection() {
-        let (_dir, runtime) = setup_enabled_cloudsync_runtime().await;
-        assert!(
-            runtime
-                .bind_cloudsync_account("user-a".to_string())
-                .await
-                .unwrap()
-        );
-        anlg_db_app::replace_cloudsync_workspace_projection(
-            runtime.pool(),
-            &anlg_db_app::CloudsyncWorkspaceProjection {
-                account_user_id: "user-a".to_string(),
-                personal_workspace_id: "user-a".to_string(),
-                workspaces: vec![anlg_db_app::CloudsyncWorkspaceProjectionEntry {
-                    id: "user-a".to_string(),
-                    owner_user_id: "user-a".to_string(),
-                    kind: "personal".to_string(),
-                    name: "Existing".to_string(),
-                    membership_id: "membership-existing".to_string(),
-                    role: "owner".to_string(),
-                    membership_created_at: "2026-07-01T01:00:00Z".to_string(),
-                    membership_updated_at: "2026-07-16T01:00:00Z".to_string(),
-                    created_at: "2026-07-01T00:00:00Z".to_string(),
-                    updated_at: "2026-07-16T00:00:00Z".to_string(),
-                }],
-            },
-        )
-        .await
-        .unwrap();
-
-        let error = runtime
-            .configure_cloudsync_token_with_projection(
-                "managed-database-id".to_string(),
-                "token".to_string(),
-                "user-a".to_string(),
-                Some(anlg_db_app::CloudsyncWorkspaceProjection {
-                    account_user_id: "user-b".to_string(),
-                    personal_workspace_id: "user-b".to_string(),
-                    workspaces: vec![anlg_db_app::CloudsyncWorkspaceProjectionEntry {
-                        id: "user-b".to_string(),
-                        owner_user_id: "user-b".to_string(),
-                        kind: "personal".to_string(),
-                        name: "Replacement".to_string(),
-                        membership_id: "membership-replacement".to_string(),
-                        role: "owner".to_string(),
-                        membership_created_at: "2026-07-01T01:00:00Z".to_string(),
-                        membership_updated_at: "2026-07-16T01:00:00Z".to_string(),
-                        created_at: "2026-07-01T00:00:00Z".to_string(),
-                        updated_at: "2026-07-16T00:00:00Z".to_string(),
-                    }],
-                }),
-                unreachable_witness("user-a"),
-            )
-            .await
-            .unwrap_err();
-
-        let workspaces: Vec<(String, String)> =
-            sqlx::query_as("SELECT id, name FROM workspaces ORDER BY id")
-                .fetch_all(runtime.pool())
-                .await
-                .unwrap();
-        assert!(matches!(
-            error,
-            crate::Error::CloudsyncWorkspace(
-                anlg_db_app::CloudsyncWorkspaceError::InvalidWorkspaceProjection
-            )
-        ));
-        assert_eq!(
-            workspaces,
-            vec![("user-a".to_string(), "Existing".to_string())]
-        );
-    }
-
-    #[tokio::test]
-    async fn token_configuration_rejects_foreign_workspace_rows() {
-        let (_dir, runtime) = setup_enabled_cloudsync_runtime().await;
-        sqlx::query("INSERT INTO sessions (id, workspace_id) VALUES ('session', 'other-user')")
-            .execute(runtime.pool())
-            .await
-            .unwrap();
-
-        assert!(
-            runtime
-                .bind_cloudsync_account("user-a".to_string())
-                .await
-                .unwrap()
-        );
-
-        let result = runtime
-            .configure_cloudsync_token(
-                "managed-database-id".to_string(),
-                "token".to_string(),
-                "user-a".to_string(),
-                unreachable_witness("user-a"),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(result, CloudsyncTokenConfigurationResult::AccountMismatch);
-        assert_eq!(
-            runtime.cloudsync_status().await.unwrap()["configured"],
-            false
-        );
-    }
-
-    #[tokio::test]
-    async fn token_configuration_rejects_an_invalid_workspace_binding() {
-        let (_dir, runtime) = setup_enabled_cloudsync_runtime().await;
-        assert!(
-            runtime
-                .bind_cloudsync_account("user-a".to_string())
-                .await
-                .unwrap()
-        );
-        sqlx::query(
-            "UPDATE app_settings
-             SET value_json = 'not-json'
-             WHERE id = 'cloudsync_workspace_binding'",
-        )
-        .execute(runtime.pool())
-        .await
-        .unwrap();
-
-        let result = runtime
-            .configure_cloudsync_token(
-                "managed-database-id".to_string(),
-                "token".to_string(),
-                "user-a".to_string(),
-                unreachable_witness("user-a"),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(result, CloudsyncTokenConfigurationResult::AccountMismatch);
-        assert_eq!(
-            runtime.cloudsync_status().await.unwrap()["configured"],
-            false
-        );
-    }
-
-    #[tokio::test]
-    async fn same_account_binding_is_idempotent() {
-        let (_dir, runtime) = setup_runtime().await;
-        assert!(
-            runtime
-                .bind_cloudsync_account("user-a".to_string())
-                .await
-                .unwrap()
-        );
-
-        assert!(
-            runtime
-                .bind_cloudsync_account("user-a".to_string())
-                .await
-                .unwrap()
-        );
-
-        assert_eq!(
-            runtime.cloudsync_status().await.unwrap()["configured"],
-            false
-        );
-    }
-
-    #[tokio::test]
-    async fn account_switch_is_rejected_and_leaves_cloudsync_suspended() {
-        let (_dir, runtime) = setup_runtime().await;
-        assert!(
-            runtime
-                .bind_cloudsync_account("user-a".to_string())
-                .await
-                .unwrap()
-        );
-
-        let bound = runtime
-            .bind_cloudsync_account("user-b".to_string())
-            .await
-            .unwrap();
-
-        assert!(!bound);
-        assert_eq!(
-            runtime.cloudsync_status().await.unwrap()["configured"],
-            false
-        );
-    }
-
-    #[tokio::test]
-    async fn invalid_account_binding_remains_an_error() {
-        let (_dir, runtime) = setup_runtime().await;
-        assert!(
-            runtime
-                .bind_cloudsync_account("user-a".to_string())
-                .await
-                .unwrap()
-        );
-
-        assert!(
-            runtime
-                .bind_cloudsync_account(" ".to_string())
-                .await
-                .is_err()
-        );
-        assert_eq!(
-            runtime.cloudsync_status().await.unwrap()["configured"],
-            false
-        );
-    }
-
-    #[tokio::test]
-    async fn invalid_sql_sends_error_through_channel() {
-        let (_dir, runtime) = setup_runtime().await;
-        let (channel, events) = capture_channel();
-
-        runtime
-            .subscribe(
-                "SELECT * FROM missing_table".to_string(),
-                vec![],
-                runtime::QueryEventChannel::new(channel),
-            )
-            .await
-            .unwrap();
-
-        let event = next_event(&events, 0).await.unwrap();
-        assert!(matches!(event, QueryEvent::Error(_)));
-    }
-}
+mod tests;
