+++ plugins/db/src/tests/mod.rs
@@ -1,0 +1,6 @@
+mod bindings;
+mod cloudsync_credentials;
+mod cloudsync_lifecycle;
+mod queries;
+mod support;
+mod transactions;
